Move BST to storage-plus
