+++ packages/storage-plus/src/lib.rs
@@ -1,4 +1,5 @@
 mod bound;
+mod bst;
 mod de;
 mod de_old;
 mod endian;
@@ -18,6 +19,7 @@ mod snapshot;
 
 #[cfg(feature = "iterator")]
 pub use bound::{Bound, Bounder, PrefixBound, RawBound};
+pub use bst::{BinarySearchTree, BinarySearchTreeIterator};
 pub use de::KeyDeserialize;
 pub use endian::Endian;
 #[cfg(feature = "iterator")]
